fix restore command bug

+++ rocketmq-redis/src/main/java/org/apache/rocketmq/redis/replicator/cmd/impl/RestoreCommand.java
@@ -24,7 +24,7 @@ public class RestoreCommand implements Command {
     private static final long serialVersionUID = 1L;
 
     private String key;
-    private int ttl;
+    private long ttl;
     private String serializedValue;
     private Boolean isReplace;
     private byte[] rawKey;
@@ -33,11 +33,11 @@ public class RestoreCommand implements Command {
     public RestoreCommand() {
     }
 
-    public RestoreCommand(String key, int ttl, String serializedValue, Boolean isReplace) {
+    public RestoreCommand(String key, long ttl, String serializedValue, Boolean isReplace) {
         this(key, ttl, serializedValue, isReplace, null, null);
     }
 
-    public RestoreCommand(String key, int ttl, String serializedValue, Boolean isReplace, byte[] rawKey,
+    public RestoreCommand(String key, long ttl, String serializedValue, Boolean isReplace, byte[] rawKey,
         byte[] rawSerializedValue) {
         this.key = key;
         this.ttl = ttl;
@@ -55,11 +55,11 @@ public class RestoreCommand implements Command {
         this.key = key;
     }
 
-    public int getTtl() {
+    public long getTtl() {
         return ttl;
     }
 
-    public void setTtl(int ttl) {
+    public void setTtl(long ttl) {
         this.ttl = ttl;
     }
